Dont publish a file if isPublishable = false
