+++ api/collection_published_test.go
@@ -1,0 +1,29 @@
+package api_test
+
+import (
+	"context"
+	"errors"
+	"io/ioutil"
+	"net/http"
+	"net/http/httptest"
+	"strings"
+	"testing"
+
+	"github.com/ONSdigital/dp-files-api/api"
+	"github.com/stretchr/testify/assert"
+)
+
+func TestPublishHandlerHandlesUnexpectedPublishingError(t *testing.T) {
+	rec := httptest.NewRecorder()
+	req := httptest.NewRequest(http.MethodPatch, "/files/ignore.txt", strings.NewReader(`{"collection_id": "asdfghjkl"}`))
+
+	h := api.HandleMarkCollectionPublished(func(ctx context.Context, collectionID string) error {
+		return errors.New("broken")
+	})
+
+	h.ServeHTTP(rec, req)
+
+	assert.Equal(t, http.StatusInternalServerError, rec.Code)
+	response, _ := ioutil.ReadAll(rec.Body)
+	assert.Contains(t, string(response), "InternalError")
+}
