+++ api/file_decrypted_test.go
@@ -1,0 +1,45 @@
+package api_test
+
+import (
+	"context"
+	"errors"
+	"io/ioutil"
+	"net/http"
+	"net/http/httptest"
+	"strings"
+	"testing"
+
+	"github.com/ONSdigital/dp-files-api/api"
+	"github.com/ONSdigital/dp-files-api/files"
+	"github.com/stretchr/testify/assert"
+)
+
+func TestDecryptedHandlerHandlesInvalidJSONContent(t *testing.T) {
+	rec := httptest.NewRecorder()
+	req := httptest.NewRequest(http.MethodPatch, "/files/file.txt", strings.NewReader("<json>invalid</json>"))
+
+	h := api.HandleMarkFileDecrypted(func(ctx context.Context, change files.FileEtagChange) error {
+		return nil
+	})
+
+	h.ServeHTTP(rec, req)
+
+	assert.Equal(t, http.StatusBadRequest, rec.Code)
+	response, _ := ioutil.ReadAll(rec.Body)
+	assert.Contains(t, string(response), "BadJsonEncoding")
+}
+
+func TestDecryptedHandlerHandlesUnexpectedPublishingError(t *testing.T) {
+	rec := httptest.NewRecorder()
+	req := httptest.NewRequest(http.MethodPatch, "/files/file.txt", strings.NewReader(`{"path": "dir/file.txt"}`))
+
+	h := api.HandleMarkFileDecrypted(func(ctx context.Context, change files.FileEtagChange) error {
+		return errors.New("broken")
+	})
+
+	h.ServeHTTP(rec, req)
+
+	assert.Equal(t, http.StatusInternalServerError, rec.Code)
+	response, _ := ioutil.ReadAll(rec.Body)
+	assert.Contains(t, string(response), "InternalError")
+}
