+++ api/get_metadata_test.go
@@ -1,0 +1,27 @@
+package api_test
+
+import (
+	"context"
+	"errors"
+	"io/ioutil"
+	"net/http"
+	"net/http/httptest"
+	"testing"
+
+	"github.com/ONSdigital/dp-files-api/api"
+	"github.com/ONSdigital/dp-files-api/files"
+	"github.com/stretchr/testify/assert"
+)
+
+func TestGetFileMetadataHandlesUnexpectedError(t *testing.T) {
+	rec := httptest.NewRecorder()
+	req := httptest.NewRequest(http.MethodGet, "/files/path.jpg", nil)
+	h := api.HandleGetFileMetadata(func(ctx context.Context, path string) (files.StoredRegisteredMetaData, error) {
+		return files.StoredRegisteredMetaData{}, errors.New("broken")
+	})
+	h.ServeHTTP(rec, req)
+
+	assert.Equal(t, http.StatusInternalServerError, rec.Code)
+	response, _ := ioutil.ReadAll(rec.Body)
+	assert.Contains(t, string(response), "InternalError")
+}
