+++ features/mark_file_as_published.feature
@@ -50,7 +50,7 @@ Feature: Mark single file as published
       | CreatedAt         | 2021-10-21T15:13:14Z                                                      |
       | LastModified      | 2021-10-21T15:14:14Z                                                      |
       | UploadCompletedAt | 2021-10-21T15:14:14Z                                                      |
-      | State             | PUBLISHED                                                                  |
+      | State             | PUBLISHED                                                                 |
       | Etag              | 123456789                                                                 |
     When the file "images/meme.jpg" is marked as published
     Then the HTTP status code should be "409"
@@ -74,3 +74,20 @@ Feature: Mark single file as published
   Scenario: The one where the file does not exists
     When the file "images/meme.jpg" is marked as published
     Then the HTTP status code should be "404"
+
+  Scenario: The one where the file is not publishable
+    Given the file upload "images/meme.jpg" has been completed with:
+      | IsPublishable     | false                                                                     |
+      | Title             | The latest Meme                                                           |
+      | SizeInBytes       | 14794                                                                     |
+      | CollectionID      | 1234-asdfg-54321-qwerty                                                   |
+      | Type              | image/jpeg                                                                |
+      | Licence           | OGL v3                                                                    |
+      | LicenceUrl        | http://www.nationalarchives.gov.uk/doc/open-government-licence/version/3/ |
+      | CreatedAt         | 2021-10-21T15:13:14Z                                                      |
+      | LastModified      | 2021-10-21T15:14:14Z                                                      |
+      | UploadCompletedAt | 2021-10-21T15:14:14Z                                                      |
+      | State             | UPLOADED                                                                  |
+      | Etag              | 123456789                                                                 |
+    When the file "images/meme.jpg" is marked as published
+    Then the HTTP status code should be "403"
