+++ store/errors.go
@@ -8,6 +8,7 @@ var (
 	ErrFileNotInCreatedState   = errors.New("file state is not in state created")
 	ErrFileNotInUploadedState  = errors.New("file state is not in state uploaded")
 	ErrFileNotInPublishedState = errors.New("file state is not in state published")
+	ErrFileIsNotPublishable    = errors.New("file is not set as publishable")
 	ErrNoFilesInCollection     = errors.New("no files found in collection")
 	ErrCollectionIDAlreadySet  = errors.New("collection ID already set")
 	ErrCollectionIDNotSet      = errors.New("collection ID not set")
